Done
Last version from me in this internship. The code is work as intended but the performance is not optimal right now. In summary the system will input the order with fixed delivery date and time, the truck data and product info and it will output the delivery plan for each day and each truck all the truck can also go back to the warehouse to recieve more item to deliver